pyproject.toml
[tool.poetry]
name = "crawjud-web"
version = "8.0.0"
description = ""
authors = ["Robotz213 <nicholasazs213@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
flask-wtf = "^1.2.2"
flask-mail = "^0.10.0"
flask-login = "^0.6.3"
flask-mysqldb = "^2.0.0"
flask-sqlalchemy = "^3.1.1"
python-dotenv = "^1.0.1"
openpyxl = "^3.1.5"
pypdf2 = "^3.0.1"
reportlab = "^4.2.5"
bcrypt = "^4.2.0"
babel = "^2.16.0"
pytz = "^2024.2"
flask-talisman = "^1.1.0"
pillow = "^11.0.0"
pandas = "^2.2.3"
deep-translator = "^1.11.4"
google-auth = "^2.36.0"
google-cloud-storage = "^2.18.2"
httpx = "^0.27.2"
psycopg2 = "^2.9.10"
flask-cloudflared = "^0.0.14"
gevent = "^24.10.3"
gevent-websocket = "^0.10.1"
pygithub = "^2.5.0"
eventlet = "^0.38.0"
clear = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black" # Garante compatibilidade com Black
line_length = 88  # O mesmo limite de linha do Black
